bot/requirements.txt
aiogram==3.13.1
groq==0.11.0
anthropic==0.36.2
supabase==2.9.1
apscheduler==3.10.4
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.10.10
pytz==2024.2
fastapi==0.115.4
uvicorn[standard]==0.32.0
